watermark_helper/requirements.txt
streamlit>=1.31.0
pdf2image>=1.17.0
Pillow>=10.4.0
numpy>=1.26.3
opencv-python-headless>=4.9.0.80
pandas>=2.1.4
openpyxl>=3.1.2
